+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Features
 
 * Sync lost messages on chatwoot
+* Set the maximum number of listeners that can be registered for events
+* Now is possible send medias with form-data
 
 ### Fixed
 
@@ -14,6 +16,19 @@
 * Fixed openai setting when send a message with chatwoot
 * Fix buildkey function in hSet and hDelete
 * Fix mexico number
+* Update baileys version
+* Update in Baileys version that fixes timeout when updating profile picture
+* Adjusts for fix timeout error on send status message
+* Chatwoot verbose logs
+* Adjusts on prisma connections
+* License terms updated
+* Fixed send message to group without no cache (local or redis)
+* Fixed startTypebot with startSession = true
+* Fixed issue of always creating a new label when saving chatwoot
+* Fixed getBase64FromMediaMessage with convertToMp4
+* Fixed bug when send message when don't have mentionsEveryOne on payload
+* Does not search message without chatwoot Message Id for reply
+* Fixed bot fallback not working on integrations
 
 # 2.1.1 (2024-09-22 10:31)